requirements.txt
flup
urllib3
iso8601
requests
flask
flask-login
flask-restful
flask-sqlalchemy
flask-marshmallow
flask-cache
marshmallow-sqlalchemy
coverage
mock
blinker
pymysql
PyJWT
psycopg2
redis